podman/qr-generator/requirements.txt
reportlab
qrcode[pil]
python-ulid
typing-extensions
